add type checking

+++ framework/models.py
@@ -14,7 +14,7 @@ class OddTs:
     sport: str | None = None
     league: str | None = None
     points: float | None = None
-    price: int | None = None
+    price: float | None = None
     is_main: bool | None = None
     is_live: bool | None = None
     timestamp: float | None = None
